FET/requirements.txt
streamlit
pandas
numpy
plotly
pillow
pytesseract
reportlab
joblib
scikit-learn
openpyxl
requests
graphviz
